Add buttonmap for DS4 connecting to macOS via BT.
Tested using DualShock 4 (2nd gen) controller connected via bluetooth
to a Macbook running macOS 11.2.3. It advertises itself as
`"DUALSHOCK 4 Wireless Controller", axes: 6, hats: 0, buttons: 14`,
not as "Wireless Controller" specified in the existing button map

https://github.com/xbmc/peripheral.joystick/blob/79d166a1519305f05ebf2db24a7fbd6bb7633083/peripheral.joystick/resources/buttonmaps/xml/cocoa/Wireless_Controller_v054C_p09CC_14b_6a.xml

So even though the vid/pid and the rest are the same, the existing
button map is not used per
https://github.com/xbmc/peripheral.joystick/blob/79d166a1519305f05ebf2db24a7fbd6bb7633083/src/storage/Device.cpp#L34

As Joystick families (#32) do not seem to be fully implemented
(adding a new entry to the joystickfamilies.xml didn't have an effect
and I don't see where it's used in the code), I solve this by
duplicating the existing button map with a different name.